Initial cli split
