+++ .gitignore
@@ -26,7 +26,6 @@ share/python-wheels/
 .installed.cfg
 *.egg
 MANIFEST
-*/_version.py
 
 # PyInstaller
 #  Usually these files are written by a python script from a template
